se agrega la ventana de confimacion de la eliminacion a clientes, producto, empleado

+++ build/web/Producto.jsp
@@ -61,7 +61,7 @@
                             <td>${producto.getEstado()}</td>
                             <td>
                                 <a class="btn btn-warning" href="Controlador?menu=Producto&accion=Editar&codigoProducto=${producto.getCodigoProducto()}">Editar</a>
-                                <a class="btn btn-danger" href="Controlador?menu=Producto&accion=Eliminar&codigoProducto=${producto.getCodigoProducto()}">Eliminar</a>
+                                <a class="btn btn-danger" onclick="return confirm('¿Quiere eliminar este registro?')" href="Controlador?menu=Producto&accion=Eliminar&codigoProducto=${producto.getCodigoProducto()}" >Eliminar</a>
                             </td>
                         </tr>
                         </c:forEach>
